refactor: tidy up OrcaClient
