+++ src/linux/Orca/isUnixSocket.ts
@@ -1,0 +1,9 @@
+import { statSync } from "fs";
+
+export function isUnixSocket(path: string): boolean {
+  try {
+    return statSync(path).isSocket();
+  } catch {
+    return false;
+  }
+}
